[Chore] #16 : S3 서비스 관련 디렉토리 변경 및 PresignedUrlVO record로 변경
